incus-osd/network: Move networking logic into its own package
