+++ incus-osd/internal/network/doc.go
@@ -1,0 +1,2 @@
+// Package network contains logic for configuring the system's network.
+package network
