+++ incus-osd/internal/network/networkd_test.go
@@ -1,4 +1,4 @@
-package systemd
+package network
 
 import (
 	"context"
