+++ incus-osd/internal/network/networkd_validate.go
@@ -1,4 +1,4 @@
-package systemd
+package network
 
 import (
 	"encoding/base64"
